LPS-52679 rename to fit jar name
