+++ modules/portal/portal-soap-extender-test/src/com/liferay/portal/soap/extender/test/handlers/SampleComponentHandler.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.soap.sample.handlers;
+package com.liferay.portal.soap.extender.test.handlers;
 
 import java.net.URL;
