+++ modules/portal/portal-soap-extender-test/src/com/liferay/portal/soap/extender/test/handlers/handlerTemplate.xsl
