+++ modules/portal/portal-soap-extender-test/test/integration/com/liferay/portal/soap/extender/test/JaxwsApiRegistrationTest.java
@@ -12,23 +12,20 @@
  * details.
  */
 
-package com.liferay.portal.soap.test;
+package com.liferay.portal.soap.extender.test;
 
-import com.liferay.portal.soap.sample.service.Greeter;
+import com.liferay.portal.soap.extender.test.service.Greeter;
 
 import java.io.IOException;
-
 import java.net.URL;
 
 import javax.xml.namespace.QName;
 import javax.xml.ws.Service;
 
 import org.arquillian.liferay.deploymentscenario.annotations.BndFile;
-
 import org.jboss.arquillian.container.test.api.RunAsClient;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.arquillian.test.api.ArquillianResource;
-
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -36,14 +33,14 @@ import org.junit.runner.RunWith;
 /**
  * @author Carlos Sierra Andrés
  */
-@BndFile("bnd-component.bnd")
+@BndFile("bnd-api.bnd")
 @RunAsClient
 @RunWith(Arquillian.class)
-public class JaxwsComponentRegistrationTest {
+public class JaxwsApiRegistrationTest {
 
 	@Test
-	public void testIsRegistered() throws IOException {
-		URL url = new URL(_url, "/o/soap/greeter?wsdl");
+	public void testGreeter() throws IOException {
+		URL url = new URL(_url, "/o/soap/greeterApi?wsdl");
 
 		QName qName = new QName(
 			"http://service.sample.soap.portal.liferay.com/",
